Remove backticks around "override"
This is a straightforward fix to eliminate any confusion that "override" might be a JavaScript keyword. Words rendered as code imply they are either a variable name or a reserved word. In this case, "override" is intended merely to convey its plain English meaning, and so should not be rendered as code.